Spectra MS backend storing data in a SQL database

The MsqlBackend is an implementation for the MsBackend() class for Spectra() objects which stores and retrieves MS data from a SQL database. New databases can be created from raw MS data files using createMsqlBackendDatabase.

MsqlBackend()

createMsqlBackendDatabase(
  dbcon,
  x = character(),
  backend = MsBackendMzR(),
  chunksize = 10L
)

# S4 method for MsBackendCached
show(object)

# S4 method for MsqlBackend
backendInitialize(object, dbcon, ...)

# S4 method for MsqlBackend
dataStorage(object)

# S4 method for MsqlBackend
[(x, i, j, ..., drop = FALSE)

# S4 method for MsqlBackend
peaksData(object)

# S4 method for MsqlBackend
intensity(object) <- value

# S4 method for MsqlBackend
mz(object) <- value

# S4 method for MsqlBackend
$(x, name) <- value

# S4 method for MsqlBackend
spectraData(object, columns = spectraVariables(object))

# S4 method for MsqlBackend
reset(object)

# S4 method for MsqlBackend
spectraNames(object)

# S4 method for MsqlBackend
spectraNames(object) <- value

Arguments

dbcon

Connection to a database.

x

For createMsqlBackendDatabase: character with the names of the raw data files from which the data should be imported. For other methods an MsqlBaackend instance.

backend

For createMsqlBackendDatabase: MS backend that can be used to import MS data from the raw files specified with parameter x.

chunksize

For createMsqlBackendDatabase: integer(1) defining the number of input that should be processed per iteration. With chunksize = 1 each file specified with x will be imported and its data inserted to the database. With chunksize = 5 data from 5 files will be imported (in parallel) and inserted to the database. Thus, higher values might result in faster database creation, but require also more memory.

object

A MsqlBackend instance.

...

For [: ignored.

i

For [: integer or logical to subset the object.

j

For [: ignored.

drop

For [: logical(1), ignored.

value

For all setter methods: replacement value.

name

For <-: character(1) with the name of the spectra variable to replace.

columns

For spectraData: character() optionally defining a subset of spectra variables that should be returned. Defaults to columns = spectraVariables(object) hence all variables are returned.

Value

See documentation of respective function.

Details

The MsqlBackend class is in principle a read-only backend but by extending the MsBackendCached() backend from the Spectra package it allows changing and adding (temporarily) spectra variables without changing the original data in the SQL database.

Creation of backend objects

SQL databases can be created and filled with MS data from raw data files using the createMsqlBackendDatabase function. Existing SQL databases (created previously with createMsqlBackendDatabase can be loaded using the conventional way to create/initialize MsBackend classes, i.e. using backendInitialize.

createMsqlBackendDatabase: create a database and fill it with MS data. Parameter dbcon is expected to be a database connection, parameter x a character vector with the file names from which to import the data. Parameter backend is used for the actual data import and defaults to backend = MsBackendMzR() hence allowing to import data from mzML, mzXML or netCDF files. Parameter chunksize allows to define the number of files (x) from which the data should be imported in one iteration. With the default chunksize = 10L data is imported from 10 files in x at the same time (if backend supports it even in parallel) and this data is then inserted into the database. Larger chunk sizes will require more memory and also larger disk space (as data import is performed through temporary files) but might eventually be faster. While data can be stored in any SQL database, at present it is suggested to use MySQL/MariaDB databases. For dbcon being a connection to a MySQL/MariaDB database, the tables will use the ARIA engine providing faster data access and will use table partitioning (using by default 10 partitions).

backendInitialize: get access and initialize a MsqlBackend object. Parameter object is supposed to be a MsqlBackend instance, created e.g. with MsqlBackend(). Parameter dbcon is expected to be a connection to a SQL database previously created with the createMsqlBackendDatabase function.

Subsetting and filtering data

MsqlBackend objects can be subsetted using the [ function. Internally, this will simply subset the integer vector of the primary keys and eventually cached data. The original data in the database is not affected by any subsetting operation. Any subsetting operation can be undone by resetting the object with the reset function. Subsetting in arbitrary order as well as index replication is supported.

In addition, MsqlBackend supports all other filtering methods available through MsBackendCached().

Accessing and modifying data

The functions listed here are specifically implemented for MsqlBackend. In addition, MsqlBackend inherits and supports all data accessor, filtering functions and data manipulation functions from MsBackendCached().

$, $<-: access or set (add) spectra variables in object. Spectra variables added or modified using the $<- are cached locally within the object (data in the database is never changed). To restore an object (i.e. drop all cached values) the reset function can be used.

dataStorage: returns a character vector same length as there are spectra in object with the name of the database containing the data.

intensity<-: not supported.

mz<-: not supported.

peaksData: returns a list with the spectras' peak data. The length of the list is equal to the number of spectra in object. Each element of the list is a matrix with columns "mz" and "intensity". For an empty spectrum, a matrix with 0 rows and two columns (named mz and intensity) is returned.

reset: restores an MsqlBackend by re-initializing it with the data from the database. Any subsetting or cached spectra variables will be lost.

spectraData: gets or general spectrum metadata. spectraData returns a DataFrame with the same number of rows as there are spectra in object. Parameter columns allows to select specific spectra variables.

spectraNames, spectraNames<-: returns a character of length equal to the number of spectra in object with the primary keys of the spectra from the database (converted to character). Replacing spectra names with spectraNames<- is not supported.

Implementation notes

Internally, the MsqlBackend class contains only the primary keys for all spectra stored in the SQL database. Keeping only these integer in memory guarantees a minimal memory footpring of the object. Still, depending of the number of spectra in the database, this integer vector might become very large. Any data access will involve SQL calls to retrieve the data from the database. By extending the MsBackendCached() object from the Spectra package, the MsqlBackend supports to (temporarily, i.e. for the duration of the R session) add or modify spectra variables. These are however stored in a data.frame within the object thus increasing the memory demand of the object.

Author

Johannes Rainer

Examples

####
## Create a new MsqlBackend database

## Define a file from which to import the data
data_file <- system.file("microtofq", "MM8.mzML", package = "msdata")

## Create a database/connection to a database
library(RSQLite)
db_file <- tempfile()
dbc <- dbConnect(SQLite(), db_file)

## Import the data from the file into the database
createMsqlBackendDatabase(dbc, data_file)
#> Importing data ... 
#> 
#> [==========================================================] 1/1 (100%) in  1s
#> 
#> Creating indices ... 
#> Done
#> [1] TRUE
dbDisconnect(dbc)

## Initialize a MsqlBackend
dbc <- dbConnect(SQLite(), db_file)
be <- backendInitialize(MsqlBackend(), dbc)

be
#> An object of class "MsqlBackend"
#> Slot "dbcon":
#> <SQLiteConnection>
#>   Path: /tmp/RtmpsLd9qy/file9c864c0cbc4
#>   Extensions: TRUE
#> 
#> Slot "spectraIds":
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54
#>  [55]  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72
#>  [73]  73  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90
#>  [91]  91  92  93  94  95  96  97  98  99 100 101 102 103 104 105 106 107 108
#> [109] 109 110 111 112 113 114 115 116 117 118 119 120 121 122 123 124 125 126
#> [127] 127 128 129 130 131 132 133 134 135 136 137 138 139 140 141 142 143 144
#> [145] 145 146 147 148 149 150 151 152 153 154 155 156 157 158 159 160 161 162
#> [163] 163 164 165 166 167 168 169 170 171 172 173 174 175 176 177 178 179 180
#> [181] 181 182 183 184 185 186 187 188 189 190 191 192 193 194 195 196 197 198
#> 
#> Slot ".tables":
#> $msms_spectrum
#>  [1] "msLevel"                  "rtime"                   
#>  [3] "acquisitionNum"           "scanIndex"               
#>  [5] "dataStorage"              "dataOrigin"              
#>  [7] "centroided"               "smoothed"                
#>  [9] "polarity"                 "precScanNum"             
#> [11] "precursorMz"              "precursorIntensity"      
#> [13] "precursorCharge"          "collisionEnergy"         
#> [15] "isolationWindowLowerMz"   "isolationWindowTargetMz" 
#> [17] "isolationWindowUpperMz"   "peaksCount"              
#> [19] "totIonCurrent"            "basePeakMZ"              
#> [21] "basePeakIntensity"        "ionisationEnergy"        
#> [23] "lowMZ"                    "highMZ"                  
#> [25] "mergedScan"               "mergedResultScanNum"     
#> [27] "mergedResultStartScanNum" "mergedResultEndScanNum"  
#> [29] "injectionTime"            "filterString"            
#> [31] "spectrumId"               "ionMobilityDriftTime"    
#> [33] "scanWindowLowerLimit"     "scanWindowUpperLimit"    
#> [35] "spectrum_id_"            
#> 
#> 
#> Slot "spectraVariables":
#>  [1] "msLevel"                  "rtime"                   
#>  [3] "acquisitionNum"           "scanIndex"               
#>  [5] "dataStorage"              "dataOrigin"              
#>  [7] "centroided"               "smoothed"                
#>  [9] "polarity"                 "precScanNum"             
#> [11] "precursorMz"              "precursorIntensity"      
#> [13] "precursorCharge"          "collisionEnergy"         
#> [15] "isolationWindowLowerMz"   "isolationWindowTargetMz" 
#> [17] "isolationWindowUpperMz"   "peaksCount"              
#> [19] "totIonCurrent"            "basePeakMZ"              
#> [21] "basePeakIntensity"        "ionisationEnergy"        
#> [23] "lowMZ"                    "highMZ"                  
#> [25] "mergedScan"               "mergedResultScanNum"     
#> [27] "mergedResultStartScanNum" "mergedResultEndScanNum"  
#> [29] "injectionTime"            "filterString"            
#> [31] "spectrumId"               "ionMobilityDriftTime"    
#> [33] "scanWindowLowerLimit"     "scanWindowUpperLimit"    
#> [35] "spectrum_id_"            
#> 
#> Slot "localData":
#> data frame with 0 columns and 198 rows
#> 
#> Slot "nspectra":
#> [1] 198
#> 
#> Slot "readonly":
#> [1] TRUE
#> 
#> Slot "version":
#> [1] "0.1"
#> 
#> Database: /tmp/RtmpsLd9qy/file9c864c0cbc4

## Original data source
head(be$dataOrigin)
#> [1] "/__w/_temp/Library/msdata/microtofq/MM8.mzML"
#> [2] "/__w/_temp/Library/msdata/microtofq/MM8.mzML"
#> [3] "/__w/_temp/Library/msdata/microtofq/MM8.mzML"
#> [4] "/__w/_temp/Library/msdata/microtofq/MM8.mzML"
#> [5] "/__w/_temp/Library/msdata/microtofq/MM8.mzML"
#> [6] "/__w/_temp/Library/msdata/microtofq/MM8.mzML"

## Data storage
head(dataStorage(be))
#> [1] "/tmp/RtmpsLd9qy/file9c864c0cbc4" "/tmp/RtmpsLd9qy/file9c864c0cbc4"
#> [3] "/tmp/RtmpsLd9qy/file9c864c0cbc4" "/tmp/RtmpsLd9qy/file9c864c0cbc4"
#> [5] "/tmp/RtmpsLd9qy/file9c864c0cbc4" "/tmp/RtmpsLd9qy/file9c864c0cbc4"

## Access all spectra data
spd <- spectraData(be)
spd
#> DataFrame with 198 rows and 37 columns
#>       msLevel     rtime acquisitionNum scanIndex                             mz
#>     <integer> <numeric>      <integer> <integer>                  <NumericList>
#> 1           1     0.486              1         1    104.554,106.996,107.966,...
#> 2           1     0.822              2         2    107.960,109.964,112.027,...
#> 3           1     1.159              3         3    107.963,112.021,113.035,...
#> 4           1     1.495              4         4    105.971,109.984,112.028,...
#> 5           1     1.832              5         5    104.474,107.000,109.968,...
#> ...       ...       ...            ...       ...                            ...
#> 194         1   65.4360            194       194    106.992,111.459,112.022,...
#> 195         1   65.7720            195       195  99.0797,109.0031,112.0232,...
#> 196         1   66.1092            196       196    102.927,108.970,112.028,...
#> 197         1   66.4458            197       197    111.058,112.026,113.029,...
#> 198         1   66.7818            198       198    111.044,112.024,112.518,...
#>           intensity            dataStorage             dataOrigin centroided
#>       <NumericList>            <character>            <character>  <integer>
#> 1      23,35,50,... /__w/_temp/Library/m.. /__w/_temp/Library/m..          1
#> 2    35, 24,140,... /__w/_temp/Library/m.. /__w/_temp/Library/m..          1
#> 3    28,179,350,... /__w/_temp/Library/m.. /__w/_temp/Library/m..          1
#> 4    25, 28,131,... /__w/_temp/Library/m.. /__w/_temp/Library/m..          1
#> 5      26,30,25,... /__w/_temp/Library/m.. /__w/_temp/Library/m..          1
#> ...             ...                    ...                    ...        ...
#> 194    34,24,86,... /__w/_temp/Library/m.. /__w/_temp/Library/m..          1
#> 195  24, 22,198,... /__w/_temp/Library/m.. /__w/_temp/Library/m..          1
#> 196    29,22,98,... /__w/_temp/Library/m.. /__w/_temp/Library/m..          1
#> 197  31,124,189,... /__w/_temp/Library/m.. /__w/_temp/Library/m..          1
#> 198    67,84,22,... /__w/_temp/Library/m.. /__w/_temp/Library/m..          1
#>      smoothed  polarity precScanNum precursorMz precursorIntensity
#>     <integer> <integer>   <integer>   <numeric>          <numeric>
#> 1          NA         1          NA          NA                 NA
#> 2          NA         1          NA          NA                 NA
#> 3          NA         1          NA          NA                 NA
#> 4          NA         1          NA          NA                 NA
#> 5          NA         1          NA          NA                 NA
#> ...       ...       ...         ...         ...                ...
#> 194        NA         1          NA          NA                 NA
#> 195        NA         1          NA          NA                 NA
#> 196        NA         1          NA          NA                 NA
#> 197        NA         1          NA          NA                 NA
#> 198        NA         1          NA          NA                 NA
#>     precursorCharge collisionEnergy isolationWindowLowerMz
#>           <integer>       <numeric>              <numeric>
#> 1                NA              NA                     NA
#> 2                NA              NA                     NA
#> 3                NA              NA                     NA
#> 4                NA              NA                     NA
#> 5                NA              NA                     NA
#> ...             ...             ...                    ...
#> 194              NA              NA                     NA
#> 195              NA              NA                     NA
#> 196              NA              NA                     NA
#> 197              NA              NA                     NA
#> 198              NA              NA                     NA
#>     isolationWindowTargetMz isolationWindowUpperMz peaksCount totIonCurrent
#>                   <numeric>              <numeric>  <integer>     <numeric>
#> 1                        NA                     NA       1743         97322
#> 2                        NA                     NA       1708         98590
#> 3                        NA                     NA       1708         96425
#> 4                        NA                     NA       1747         97144
#> 5                        NA                     NA       1730         94631
#> ...                     ...                    ...        ...           ...
#> 194                      NA                     NA       2267        142472
#> 195                      NA                     NA       2488        167102
#> 196                      NA                     NA       2324        137565
#> 197                      NA                     NA       2105        113086
#> 198                      NA                     NA       1996         97257
#>     basePeakMZ basePeakIntensity ionisationEnergy     lowMZ    highMZ
#>      <numeric>         <numeric>        <numeric> <numeric> <numeric>
#> 1      144.050              7250                0   104.554  1004.470
#> 2      144.050              7064                0   107.960  1002.799
#> 3      144.050              7124                0   107.963   993.959
#> 4      144.051              7067                0   105.971   986.756
#> 5      144.051              6891                0   104.474  1003.469
#> ...        ...               ...              ...       ...       ...
#> 194    144.051              4615                0  106.9921  1003.828
#> 195    144.051              3812                0   99.0797  1003.270
#> 196    144.050              3526                0  102.9272  1000.594
#> 197    144.050              3902                0  111.0582   984.228
#> 198    144.051              3587                0  111.0437  1000.851
#>     mergedScan mergedResultScanNum mergedResultStartScanNum
#>      <integer>           <integer>                <integer>
#> 1           NA                  NA                       NA
#> 2           NA                  NA                       NA
#> 3           NA                  NA                       NA
#> 4           NA                  NA                       NA
#> 5           NA                  NA                       NA
#> ...        ...                 ...                      ...
#> 194         NA                  NA                       NA
#> 195         NA                  NA                       NA
#> 196         NA                  NA                       NA
#> 197         NA                  NA                       NA
#> 198         NA                  NA                       NA
#>     mergedResultEndScanNum injectionTime filterString  spectrumId
#>                  <integer>     <numeric>  <character> <character>
#> 1                       NA             0           NA      scan=1
#> 2                       NA             0           NA      scan=2
#> 3                       NA             0           NA      scan=3
#> 4                       NA             0           NA      scan=4
#> 5                       NA             0           NA      scan=5
#> ...                    ...           ...          ...         ...
#> 194                     NA             0           NA    scan=194
#> 195                     NA             0           NA    scan=195
#> 196                     NA             0           NA    scan=196
#> 197                     NA             0           NA    scan=197
#> 198                     NA             0           NA    scan=198
#>     ionMobilityDriftTime scanWindowLowerLimit scanWindowUpperLimit spectrum_id_
#>                <numeric>            <numeric>            <numeric>    <integer>
#> 1                     NA                   NA                   NA            1
#> 2                     NA                   NA                   NA            2
#> 3                     NA                   NA                   NA            3
#> 4                     NA                   NA                   NA            4
#> 5                     NA                   NA                   NA            5
#> ...                  ...                  ...                  ...          ...
#> 194                   NA                   NA                   NA          194
#> 195                   NA                   NA                   NA          195
#> 196                   NA                   NA                   NA          196
#> 197                   NA                   NA                   NA          197
#> 198                   NA                   NA                   NA          198

## Available variables
spectraVariables(be)
#>  [1] "msLevel"                  "rtime"                   
#>  [3] "acquisitionNum"           "scanIndex"               
#>  [5] "mz"                       "intensity"               
#>  [7] "dataStorage"              "dataOrigin"              
#>  [9] "centroided"               "smoothed"                
#> [11] "polarity"                 "precScanNum"             
#> [13] "precursorMz"              "precursorIntensity"      
#> [15] "precursorCharge"          "collisionEnergy"         
#> [17] "isolationWindowLowerMz"   "isolationWindowTargetMz" 
#> [19] "isolationWindowUpperMz"   "peaksCount"              
#> [21] "totIonCurrent"            "basePeakMZ"              
#> [23] "basePeakIntensity"        "ionisationEnergy"        
#> [25] "lowMZ"                    "highMZ"                  
#> [27] "mergedScan"               "mergedResultScanNum"     
#> [29] "mergedResultStartScanNum" "mergedResultEndScanNum"  
#> [31] "injectionTime"            "filterString"            
#> [33] "spectrumId"               "ionMobilityDriftTime"    
#> [35] "scanWindowLowerLimit"     "scanWindowUpperLimit"    
#> [37] "spectrum_id_"            

## Access mz values
mz(be)
#> NumericList of length 198
#> [[1]] 104.553733825684 106.996170043945 ... 1002.67761230469 1004.47033691406
#> [[2]] 107.959617614746 109.964324951172 ... 991.343811035156 1002.79937744141
#> [[3]] 107.962753295898 112.021392822266 ... 964.238586425781 993.959167480469
#> [[4]] 105.970863342285 109.984230041504 ... 979.084838867188 986.755981445312
#> [[5]] 104.473876953125 106.999610900879 ... 1003.185546875 1003.46936035156
#> [[6]] 102.036674499512 107.963928222656 ... 993.781127929688 994.755920410156
#> [[7]] 104.272621154785 112.024925231934 ... 996.248046875 1001.64440917969
#> [[8]] 98.9297180175781 106.506416320801 ... 991.034484863281 997.045837402344
#> [[9]] 107.966445922852 109.385284423828 ... 995.031494140625 995.600463867188
#> [[10]] 108.965042114258 109.974197387695 ... 982.596313476562 989.777404785156
#> ...
#> <188 more elements>

## Subset the object to spectra in arbitrary order
be_sub <- be[c(5, 1, 1, 2, 4, 100)]
be_sub
#> An object of class "MsqlBackend"
#> Slot "dbcon":
#> <SQLiteConnection>
#>   Path: /tmp/RtmpsLd9qy/file9c864c0cbc4
#>   Extensions: TRUE
#> 
#> Slot "spectraIds":
#> [1]   5   1   1   2   4 100
#> 
#> Slot ".tables":
#> $msms_spectrum
#>  [1] "msLevel"                  "rtime"                   
#>  [3] "acquisitionNum"           "scanIndex"               
#>  [5] "dataStorage"              "dataOrigin"              
#>  [7] "centroided"               "smoothed"                
#>  [9] "polarity"                 "precScanNum"             
#> [11] "precursorMz"              "precursorIntensity"      
#> [13] "precursorCharge"          "collisionEnergy"         
#> [15] "isolationWindowLowerMz"   "isolationWindowTargetMz" 
#> [17] "isolationWindowUpperMz"   "peaksCount"              
#> [19] "totIonCurrent"            "basePeakMZ"              
#> [21] "basePeakIntensity"        "ionisationEnergy"        
#> [23] "lowMZ"                    "highMZ"                  
#> [25] "mergedScan"               "mergedResultScanNum"     
#> [27] "mergedResultStartScanNum" "mergedResultEndScanNum"  
#> [29] "injectionTime"            "filterString"            
#> [31] "spectrumId"               "ionMobilityDriftTime"    
#> [33] "scanWindowLowerLimit"     "scanWindowUpperLimit"    
#> [35] "spectrum_id_"            
#> 
#> 
#> Slot "spectraVariables":
#>  [1] "msLevel"                  "rtime"                   
#>  [3] "acquisitionNum"           "scanIndex"               
#>  [5] "dataStorage"              "dataOrigin"              
#>  [7] "centroided"               "smoothed"                
#>  [9] "polarity"                 "precScanNum"             
#> [11] "precursorMz"              "precursorIntensity"      
#> [13] "precursorCharge"          "collisionEnergy"         
#> [15] "isolationWindowLowerMz"   "isolationWindowTargetMz" 
#> [17] "isolationWindowUpperMz"   "peaksCount"              
#> [19] "totIonCurrent"            "basePeakMZ"              
#> [21] "basePeakIntensity"        "ionisationEnergy"        
#> [23] "lowMZ"                    "highMZ"                  
#> [25] "mergedScan"               "mergedResultScanNum"     
#> [27] "mergedResultStartScanNum" "mergedResultEndScanNum"  
#> [29] "injectionTime"            "filterString"            
#> [31] "spectrumId"               "ionMobilityDriftTime"    
#> [33] "scanWindowLowerLimit"     "scanWindowUpperLimit"    
#> [35] "spectrum_id_"            
#> 
#> Slot "localData":
#> data frame with 0 columns and 6 rows
#> 
#> Slot "nspectra":
#> [1] 6
#> 
#> Slot "readonly":
#> [1] TRUE
#> 
#> Slot "version":
#> [1] "0.1"
#> 
#> Database: /tmp/RtmpsLd9qy/file9c864c0cbc4

## The internal spectrum IDs (primary keys from the database)
be_sub$spectrum_id_
#> [1]   5   1   1   2   4 100

## Add additional spectra variables
be_sub$new_variable <- "B"

## This variable is *cached* locally within the object (not inserted into the
## database)
be_sub$new_variable
#> [1] "B" "B" "B" "B" "B" "B"